OK, we can now run the MaraDNS tests
We now have a wrapper which converts a function from MaraDNS’s API
for HalfSip13 to the API used by the reference code.

We also have code which generates the same output as the code which
will be the Deadwood HalfSipHash test.

The reason for this is because the official vectors are limited, and
this allows me to make more HalfSip test vectors.